feature(bento-mega-menu): moved useMutationObserver to common location
